# Implement prepared schemas for faster decoding