Preparing a schema returns an immutable schema so that we can
implement speedups conferred by knowing the value will not change.

Presently, only decoding speedups are implemented but this leaves
plenty of room for far more speedups in the future.

---

This is the first step of what I was talking about in [this comment](https://github.com/elodina/go-avro/issues/48#issuecomment-142037140)

Decoding speedups are primarily achieved by further memoizing the struct field lookup stuff (and enums, too) in a lock-free manner. The very first decode to a given struct type takes double the time because it's analyzing it, but repeated decoding yields noticeable speedups: 
33% on a single-threaded benchmark and an even more noticeable 40% on a multi-threaded bench. (x86-64, darwin)

```
single-thread:
BenchmarkSpecificDatumReader_complex            	  500000	      3008 ns/op	     304 B/op	      12 allocs/op
BenchmarkSpecificDatumReader_complex_prepared   	 1000000	      1989 ns/op	     304 B/op	      12 allocs/op
BenchmarkSpecificDatumReader_hugeval            	  500000	      3173 ns/op	     304 B/op	      12 allocs/op
BenchmarkSpecificDatumReader_hugeval_prepared   	 1000000	      2055 ns/op	     304 B/op	      12 allocs/op


parallel:

BenchmarkSpecificDatumReader_complex-4            	 1000000	      1807 ns/op	     304 B/op	      12 allocs/op
BenchmarkSpecificDatumReader_complex_prepared-4   	 1000000	      1067 ns/op	     304 B/op	      12 allocs/op
BenchmarkSpecificDatumReader_hugeval-4            	 1000000	      1710 ns/op	     304 B/op	      12 allocs/op
BenchmarkSpecificDatumReader_hugeval_prepared-4   	 1000000	      1092 ns/op	     304 B/op	      12 allocs/op
```