@crast 2016-03-14
@serejja  Do you still want support for Go 1.1 and 1.2 ? If you do it's going to take more code because those versions don't have sync.Pool