@serejja 2016-03-14
Thanks @crast! I'm going to drop Go 1.1 and 1.2 support in CI now (and also add 1.6). Nobody uses 1.1 and 1.2 now anyway.

I'm also working on one other thing that is somewhat intertwined with the changes you made, which is in [this](https://github.com/elodina/go-avro/tree/codegen-reader-writer) branch. Basically, go-avro codegen will support generating Write() and Read() functions to avoid using reflection at all. Similar approach is used in `json` package (e.g. `MarshalJSON`, `UnmarshalJSON`) and could be an option to speed things up in a lot of cases too.

Any comments and thoughts appreciated.
Thanks for your contributions!