@crast 2016-03-14
@serejja Awesome. Yeah that will produce noticeable speedups.