@crast 2016-03-14
I daresay that the codegen will make it an order of magnitude faster, because at this point the majority of time is spent in `reflect`. I remember when single-threaded decoding hugeval [was 23000ns on my machine](https://github.com/elodina/go-avro/pull/45), now it's down to 2000ns, will be curious to see what you get out of the codegen approach.